Split models to separate folders
