+++ spec/services/session_manipulation/add_vehicle_details_spec.rb
@@ -7,7 +7,7 @@ describe SessionManipulation::AddVehicleDetails do
     described_class.call(params: params, session: session)
   end
 
-  let(:params) { ChargeableFleet.new(data).vehicle_list }
+  let(:params) { ManageVehicles::ChargeableFleet.new(data).vehicle_list }
   let(:data) { read_response('chargeable_vehicles.json') }
   let(:session) { {} }
   let(:first_vehicle) { params.first }
